+++ nutrition/Model/MealIngredient.swift
@@ -51,12 +51,12 @@ class MealIngredientMgr: ObservableObject {
     }
 
     func create(name: String,
-                defaultAmount: Double,
-                amount: Double,
+                defaultAmount: Float,
+                amount: Float,
                 consumptionUnit: Unit,
                 compensationExists: Bool = false,
                 compensationCreated: Bool = false,
-                compensationInitialAmount: Double = 0,
+                compensationInitialAmount: Float = 0,
                 compensationInitialState: Bool = true) {
 
         let mealIngredient = MealIngredient(name: name,
@@ -127,13 +127,13 @@ class MealIngredientMgr: ObservableObject {
         return inactiveIngredients.count > 0
     }
 
-    func addMacros(name: String, calories: Double, fat: Double, fiber: Double, netcarbs: Double, protein: Double) {
+    func addMacros(name: String, calories: Float, fat: Float, fiber: Float, netcarbs: Float, protein: Float) {
         if let index = mealIngredients.firstIndex(where: { $0.name == name }) {
             mealIngredients[index] = mealIngredients[index].addMacros(calories: calories, fat: fat, fiber: fiber, netcarbs: netcarbs, protein: protein)
         }
     }
 
-    func adjust(name: String, amount: Double, consumptionUnit: Unit) {
+    func adjust(name: String, amount: Float, consumptionUnit: Unit) {
         if let index = mealIngredients.firstIndex(where: { $0.name == name && !$0.active }) {
             mealIngredients[index] = mealIngredients[index].adjust(amount: amount, active: false)
         } else if let index = mealIngredients.firstIndex(where: { $0.name == name }) {
@@ -200,36 +200,36 @@ struct MealIngredient: Codable, Identifiable {
 
     var name: String
 
-    var defaultAmount: Double
-    var amount: Double
+    var defaultAmount: Float
+    var amount: Float
     var consumptionUnit: Unit
 
-    var calories: Double
-    var fat: Double
-    var fiber: Double
-    var netcarbs: Double
-    var protein: Double
+    var calories: Float
+    var fat: Float
+    var fiber: Float
+    var netcarbs: Float
+    var protein: Float
 
     var compensationExists: Bool
     var compensationCreated: Bool
-    var compensationInitialAmount: Double
+    var compensationInitialAmount: Float
     var compensationInitialState: Bool
 
     var active: Bool
 
     init(id: String = UUID().uuidString,
          name: String,
-         defaultAmount: Double,
-         amount: Double,
+         defaultAmount: Float,
+         amount: Float,
          consumptionUnit: Unit = Unit.gram,
-         calories: Double = 0,
-         fat: Double = 0,
-         fiber: Double = 0,
-         netcarbs: Double = 0,
-         protein: Double = 0,
+         calories: Float = 0,
+         fat: Float = 0,
+         fiber: Float = 0,
+         netcarbs: Float = 0,
+         protein: Float = 0,
          compensationExists: Bool = false,
          compensationCreated: Bool = false,
-         compensationInitialAmount: Double = 0,
+         compensationInitialAmount: Float = 0,
          compensationInitialState: Bool = true,
          active: Bool = true) {
 
@@ -263,11 +263,11 @@ struct MealIngredient: Codable, Identifiable {
         return MealIngredient(id: id, name: name, defaultAmount: defaultAmount, amount: amount, consumptionUnit: consumptionUnit, calories: 0, fat: 0, fiber: 0, netcarbs: 0, protein: 0, compensationExists: compensationExists, compensationCreated: compensationCreated, compensationInitialAmount: compensationInitialAmount, compensationInitialState: compensationInitialState, active: active);
     }
 
-    func addMacros(calories: Double, fat: Double, fiber: Double, netcarbs: Double, protein: Double) -> MealIngredient {
+    func addMacros(calories: Float, fat: Float, fiber: Float, netcarbs: Float, protein: Float) -> MealIngredient {
         return MealIngredient(id: id, name: name, defaultAmount: defaultAmount, amount: amount, consumptionUnit: consumptionUnit, calories: self.calories + calories, fat: self.fat + fat, fiber: self.fiber + fiber, netcarbs: self.netcarbs + netcarbs, protein: self.protein + protein, compensationExists: compensationExists, compensationCreated: compensationCreated, compensationInitialAmount: compensationInitialAmount, compensationInitialState: compensationInitialState, active: active);
     }
 
-    func adjust(amount: Double, active: Bool = true) -> MealIngredient {
+    func adjust(amount: Float, active: Bool = true) -> MealIngredient {
         if !compensationExists && !active {
             print("  Adjusting (activating - and storing compensation) " + name + " \(amount)")
             return MealIngredient(id: self.id, name: self.name, defaultAmount: self.defaultAmount, amount: amount, consumptionUnit: self.consumptionUnit, calories: calories, fat: fat, fiber: fiber, netcarbs: netcarbs, protein: protein, compensationExists: true, compensationCreated: false, compensationInitialAmount: self.amount, compensationInitialState: self.active, active: true)
